Update Platfrom and Test
