+++ src/test/java/org/example/databaseClient/databaseClientTest.java
@@ -4,6 +4,8 @@ import com.mongodb.client.MongoCollection;
 import org.bson.Document;
 import org.example.model.Activity.Activity;
 import org.example.model.User.User;
+import org.example.platform.IUserPlatform;
+import org.example.platform.UserPlatform;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -20,11 +22,13 @@ public class databaseClientTest {
     private static final String ACTIVITY_COLLECTION_NAME = "activity";
 
     private databaseClient client;
+    private IUserPlatform userPlatform;
 
     @BeforeEach
     public void setup() {
         client = new databaseClient();
         client.init();
+        userPlatform = new UserPlatform(client.getUserCollection());
     }
 
     @AfterEach
@@ -42,7 +46,7 @@ public class databaseClientTest {
         User user = new User("John", "Doe", new Date(), "Male");
 
         // Register the user
-        client.register(user);
+        userPlatform.register(user);
 
         // Retrieve the user document from the database
         MongoCollection<Document> userCollection = client.getMongoClient().getDatabase(DATABASE_NAME).getCollection(USER_COLLECTION_NAME);
@@ -61,13 +65,13 @@ public class databaseClientTest {
         User user = new User("John", "Doe", new Date(), "Male");
 
         // Register the user
-        client.register(user);
+        userPlatform.register(user);
 
         // Create an activity
         Activity activity = new Activity("Running", 30, 8, new Date());
 
         // Create the activity for the user
-        client.addActivityToUser(user, activity);
+        userPlatform.addActivityToUser(user, activity);
 
         // Verify that the activity is associated with the user
         // You may need to modify this assertion based on your actual implementation
